requirements.txt
# --------- base-tool --------- #
sphinx==8.1.3
PyYAML
sphinx-rtd-theme==3.0.2
pytest==8.3.5
pytest-cov==6.0.0
pyrootutils==1.0.4
pre-commit==4.1.0
twine==6.1.0
wheel>=0.45.1
coverage==7.6.12
furo==2024.8.6
myst-parser==4.0.1
pytest-mock==3.14.0
pendulum==3.0.0
python-dotenv==1.0.1
cookiecutter==2.6.0
setuptools

# --------- date --------- #
openpyxl==3.1.5
attrs>=25.1.0,<26
duckdb-engine
SQLAlchemy

# --------- path --------- #
pathlib==1.0.1


# --------- cli --------- #
click
# --------- log --------- #
prettytable
websocket-client
sounddevice
webrtcvad
PyAudio
numpy

# --------- llm --------- #
openai


# --------- ai --------- #
funasr
ray
mem0ai
langchain_openai
langchain
langchain-mcp-adapters
langgraph
agno
duckdb
fastmcp
httpx
psutil

# --------- server --------- #

fastapi
uvicorn
websockets
numpy
scipy
python-dotenv
webrtcvad
silero-vad 
pysilero-vad
apscheduler